Zmiany w village service - znajdywanie wioski gracza po jego id, w use service mozliwosc pobrania info o aktualnie zalogowanym graczu, nowy kontroler attack, zmiany na mapie, right click, odwiedz lub atakuj

+++ Strona/projekt_plemiona/src/main/java/com/example/projekt_plemiona/models/VillageUnitsId.java
@@ -13,5 +13,5 @@ import lombok.AllArgsConstructor;
 public class VillageUnitsId implements Serializable {
 
     private Long villageId;
-    private int unitTypeId;
+    private Long unitTypeId;
 }
